gemini-langchain/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain-community
langchain
PyPDF2 
faiss-cpu
langchain_google_genai